+++ CHANGES.md
@@ -1,5 +1,6 @@
 #4.0.0.beta3 2014-10-07
  - 2014-10-14 Fixed bug#10579 - Error with &-sign in an eMail address.
+ - 2014-10-14 Fixed bug#[10621](http://bugs.otrs.org/show_bug.cgi?id=10621) - Queues are false sorted in Ticket Zoom select box.
  - 2014-10-13 Changed default skin for CKEditor to bootstrapck, http://ckeditor.com/addon/bootstrapck.
  - 2014-10-07 Fixed support data collection hangs in some mod_perl environments.
  - 2014-10-07 Added the possibility to restrict customer self registration by email address whitelist or blacklist.